backend/pyproject.toml
[tool.poetry]
name = "weddinghousedashboard"
version = "0.1.0"
description = ""
authors = ["differentmonster <differentcodemonster@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
# python = ">=3.6,<3.9"
asgiref = "^3.2.10"
autopep8 = "^1.5.3"
django = "^3.0.7"
pycodestyle = "^2.6.0"
pytz = "^2020.1"
sqlparse = "^0.3.1"
Unipath = "^1.1"
dj-database-url = "^0.5.0"
gunicorn = "^20.0.4"
whitenoise = "^5.1.0"
django-rest-framework = "^0.1.0"
pillow = "^7.1.2"
django-pdb = "^0.6.2"
django-environ = "^0.4.5"
psycopg2 = "^2.8.5"
django-cors-headers = "^3.5.0"
django-countries = "^6.1.3"
djangorestframework-simplejwt = "^4.4.0"
pytest = "^6.2.1"
pytest-django = "^4.1.0"
pytest-cov = "^2.11.1"
djoser = "^2.1.0"

[tool.poetry.dev-dependencies]

[build-system]
# Should be the same as `$POETRY_VERSION`:
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
